more: extract client component into /modules folder
