+++ apps/web/app/future/more/page.tsx
@@ -1,4 +1,5 @@
-import Page from "@pages/more";
 import { WithLayout } from "app/layoutHOC";
+
+import Page from "~/more/more-page-view";
 
 export default WithLayout({ getLayout: null, Page })<"P">;
